refactor(run): move timing and grouping helpers into run/ modules
run.ts held three unrelated concerns. The mechanical ones now live beside
tests-in-browser.ts, leaving run.ts to the two modes and their HTML caching:

- run/timings.ts   readTimingCache, computeFileTimes, persistTimings,
                   printFileTimings — the cache that feeds LPT packing
- run/grouping.ts  applyWatchLineTargets, resolveTargetedFiles,
                   splitIntoGroups, TargetedFile — how a run becomes groups

Also fixed while moving:

- the LPT bin-packing comment had drifted off splitIntoGroups and was stranded
  above applyWatchLineTargets' JSDoc, so the algorithm it explains was
  undocumented and an unrelated function carried a stray comment. Reunited.
- run.ts exported buildCachedContent with no consumers, and re-exported
  readTimingCache/computeFileTimes purely so a test could reach them; the test
  now imports run/timings.ts directly and the export line is gone.
- `path` became unused in run.ts once resolveTargetedFiles moved out, and
  persistTimings/printFileTimings/splitIntoGroups needed JSDoc as new exports.

Pure moves — no behavior change.
