+++ lib/commands/run.ts
@@ -9,7 +9,6 @@ import {
 } from '../setup/web-server.ts';
 import { openOutputInBrowser } from '../utils/open-output-in-browser.ts';
 import fs from 'node:fs/promises';
-import path from 'node:path';
 import { normalize } from 'node:path';
 import { availableParallelism } from 'node:os';
 // node:timers returns Timer objects with .unref()/.ref() in both Node and Deno.
@@ -43,7 +42,13 @@ import {
 } from '../utils/failure-cache.ts';
 import { writeCoverageReport } from '../coverage/report.ts';
 import { isFilteredRun, describeFilter } from '../selection/filter-query.ts';
-import { resolveLineTargets } from '../selection/line-targets.ts';
+import {
+  readTimingCache,
+  computeFileTimes,
+  persistTimings,
+  printFileTimings,
+} from './run/timings.ts';
+import { applyWatchLineTargets, resolveTargetedFiles, splitIntoGroups } from './run/grouping.ts';
 import type { QUnitSelector } from '../selection/line-targets.ts';
 import type { Config, CachedContent } from '../types.ts';
 
@@ -653,7 +658,6 @@ async function runConcurrentMode(
   });
 }
 
-export { readTimingCache, computeFileTimes, buildCachedContent };
 export { run as default };
 
 /**
@@ -729,159 +733,6 @@ async function addCachedContentMainHTML(
   }
 
   return cachedContent;
-}
-
-/** Reads `tmp/test-timings.json` from projectRoot; returns `{}` on any error or invalid content. */
-async function readTimingCache(projectRoot: string): Promise<Record<string, number>> {
-  try {
-    const parsed = JSON.parse(await fs.readFile(`${projectRoot}/tmp/test-timings.json`, 'utf8'));
-    return typeof parsed === 'object' && parsed !== null && !Array.isArray(parsed) ? parsed : {};
-  } catch {
-    return {};
-  }
-}
-
-/** Distributes each group's wall-clock ms to its files proportionally by LPT weight. */
-function computeFileTimes(
-  groups: string[][],
-  weights: Map<string, number>,
-  wallTimes: Map<number, number>,
-): Map<string, number> {
-  const result = new Map<string, number>();
-  groups.forEach((group, i) => {
-    const wallMs = wallTimes.get(i);
-    if (wallMs === undefined) return;
-    const total = group.reduce((sum, f) => sum + (weights.get(f) ?? 0), 0);
-    group.forEach((f) =>
-      result.set(f, total > 0 ? wallMs * ((weights.get(f) ?? 0) / total) : wallMs / group.length),
-    );
-  });
-  return result;
-}
-
-async function persistTimings(fileTimes: Map<string, number>, projectRoot: string): Promise<void> {
-  await fs.writeFile(
-    `${projectRoot}/tmp/test-timings.json`,
-    JSON.stringify(Object.fromEntries(fileTimes), null, 2),
-  );
-}
-
-function printFileTimings(fileTimes: Map<string, number>, projectRoot: string): void {
-  if (fileTimes.size === 0) return;
-  const lines = [...fileTimes.entries()]
-    .sort(([, a], [, b]) => b - a)
-    .map(([f, ms]) => `#   ${ms.toFixed(0)}ms  ${f.replace(`${projectRoot}/`, '')}`);
-  process.stdout.write(`# File execution times:\n${lines.join('\n')}\n`);
-}
-
-// LPT (Longest Processing Time first) bin-packing: sort files by estimated time descending,
-// then assign each to the group with the smallest current total. Uses cached per-file timings
-// when available; falls back to file size scaled by msPerByte for unknown files.
-/**
- * Watch-mode line targets: narrow fsTree to the targeted files and apply their selectors for the
- * whole session.
- *
- * Watch is one page with one QUnit config, so a page-global selector set would filter every OTHER
- * file's tests down to nothing on the next save — the per-file scoping concurrent mode gets from
- * one-group-per-file has nowhere to live here. Narrowing fsTree keeps the selectors true for
- * everything loaded, at the cost of dropping untargeted inputs; those are named rather than
- * silently watched-but-never-run. `qa` clears the selectors to run everything still watched.
- */
-async function applyWatchLineTargets(config: Config): Promise<void> {
-  const allFiles = Object.keys(config.fsTree);
-  const targets = await resolveTargetedFiles(config, allFiles);
-  if (targets.length === 0) return;
-
-  const targetedPaths = new Set(targets.map((target) => target.file));
-  const dropped = allFiles.filter((file) => !targetedPaths.has(file));
-  config.fsTree = Object.fromEntries([...targetedPaths].map((file) => [file, config.fsTree[file]]));
-  config._qunitSelectors = targets.flatMap((target) => target.selectors);
-  if (dropped.length > 0) {
-    console.log(
-      '#',
-      blue(
-        `qunitx: --watch with a line target runs only the targeted file${targetedPaths.size === 1 ? '' : 's'} — ${dropped.length} other file${dropped.length === 1 ? '' : 's'} excluded from this session`,
-      ),
-    );
-  }
-  console.log('#', blue(`qunitx: press "qa" to run every test in the watched file(s)`));
-}
-
-/** A file whose run is narrowed by `file#34` targets, with the selectors that scope it. */
-interface TargetedFile {
-  file: string;
-  selectors: QUnitSelector[];
-}
-
-/**
- * Resolves each `file#34` input into the selectors for that file, dropping targets whose file is
- * no longer in the run (a glob, `--changed` or `--only-failed` may have filtered it out) and
- * those that resolved to nothing — both fall back to running the file whole, which is what a
- * null `selectors` means. Every warning is surfaced; a line target that quietly did not narrow
- * is worse than one that says so.
- */
-async function resolveTargetedFiles(config: Config, allFiles: string[]): Promise<TargetedFile[]> {
-  const present = new Set(allFiles);
-  const entries = Object.entries(config.lineTargets ?? {}).filter(([file]) => present.has(file));
-  const resolved = await Promise.all(
-    entries.map(async ([file, lines]) => {
-      const { selectors, warnings } = await resolveLineTargets(
-        file,
-        lines,
-        // Forward slashes in the warning regardless of OS — it echoes the `path#line` the user
-        // typed, and they typed '/'. path.relative yields '\' on Windows.
-        path.relative(config.projectRoot, file).replaceAll('\\', '/'),
-      );
-
-      return { file, selectors, warnings };
-    }),
-  );
-  // Logged after the parallel resolve so warnings print in input order, not resolution order.
-  resolved.forEach((entry) => {
-    entry.warnings.forEach((warning) => console.log('#', blue(`qunitx: ${warning}`)));
-  });
-
-  // Null selectors mean the target degraded to "run the whole file" — there is nothing to scope,
-  // so it is not a targeted file and falls through to the untargeted pool.
-  return resolved
-    .filter((entry) => entry.selectors !== null)
-    .map(({ file, selectors }) => ({ file, selectors: selectors! }));
-}
-
-async function splitIntoGroups(
-  files: string[],
-  groupCount: number,
-  timings: Record<string, number>,
-): Promise<{ groups: string[][]; weights: Map<string, number> }> {
-  const sizes = await Promise.all(
-    files.map((f) =>
-      timings[f] > 0
-        ? Promise.resolve(0)
-        : fs
-            .stat(f)
-            .then((s) => s.size)
-            .catch(() => 0),
-    ),
-  );
-  const knownRates = files
-    .map((f, i) => ({ ms: timings[f], size: sizes[i] }))
-    .filter(({ ms, size }) => ms > 0 && size > 0);
-  const msPerByte =
-    knownRates.length > 0
-      ? knownRates.reduce((sum, { ms, size }) => sum + ms / size, 0) / knownRates.length
-      : 1;
-  const weights = new Map(
-    files.map((f, i) => [f, timings[f] > 0 ? timings[f] : sizes[i] * msPerByte]),
-  );
-  const buckets = Array.from({ length: groupCount }, () => ({ files: [] as string[], total: 0 }));
-  [...files]
-    .sort((a, b) => (weights.get(b) ?? 0) - (weights.get(a) ?? 0))
-    .forEach((f) => {
-      const min = buckets.reduce((m, _, i) => (buckets[i].total < buckets[m].total ? i : m), 0);
-      buckets[min].files.push(f);
-      buckets[min].total += weights.get(f) ?? 0;
-    });
-  return { groups: buckets.filter((b) => b.files.length > 0).map((b) => b.files), weights };
 }
 
 function logWatcherAndKeyboardShortcutInfo(config: Config, _server: unknown): void {
